move files into the right folder
