+++ libraries/utilities/index.ts
@@ -4,8 +4,8 @@ export { dataAttributes } from './functions/dataAttributes'
 export { deburr } from './functions/deburr'
 export { decode, encode } from './functions/encode'
 export { deepEquals } from './functions/deepEquals'
-export { formatBytes } from './formatBytes'
-export { formatMilliseconds } from './formatMilliseconds'
+export { formatBytes } from './functions/formatBytes'
+export { formatMilliseconds } from './functions/formatMilliseconds'
 export { get } from './functions/get'
 export { hash } from './functions/hash'
 export { HttpVerb } from './types/HttpVerb'
